[update] 一部関数の記述を変更

+++ src/MoveLineEffect.cpp
@@ -37,8 +37,8 @@ void MoveLineEffect::initialize()
 	const static float INSTANCE_RADIUS = 60.0f;
 
 	const auto cameraPosX = mGameObjectList->findGameObject(GameObjectTag::Player)->transform().pos.x;
-	const int angleDegree = Random::RandomInt(360);
-	const float angleRadian = angleDegree * 3.141592653589793 / 180.0f;
+	const float angleDegree = (float)Random::RandomInt(360);
+	const float angleRadian = angleDegree * 3.141592653589793f / 180.0f;
 	mTransform.pos.x = cameraPosX + INSTANCE_RADIUS * cos(angleRadian);
 	mTransform.pos.y = INSTANCE_RADIUS * sin(angleRadian);
 	mTransform.pos.z = 600.0f;
